requirements.txt
asgiref==3.2.3
autopep8==1.5
django-cors-headers==3.2.1
django-templated-mail==1.1.1
django-webpack-loader==0.6.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
mysqlclient==1.4.6
Pillow==7.0.0
pycodestyle==2.5.0
PyJWT==1.7.1
pytz==2019.3
SQLAlchemy==1.3.13
sqlparse==0.3.0
sqlservice==1.2.1
